Module2/order_api/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pika==1.3.2
mysql-connector-python==8.3.0
pydantic==2.7.1
